Modified logging statements to use PRI formatting.

+++ components/ALU/Register.cpp
@@ -1,5 +1,6 @@
 #include <cstdlib>
 #include <string>
+#include <inttypes.h>
 
 #include "Instructions.hpp"
 #include "Register.hpp"
@@ -171,7 +172,7 @@ string Register::Disassemble() const
     }
     else
     {
-        snprintf(buffer, sizeof(buffer), "%s %d", Instructions::Mnemonic(opcode), lineNumber);
+        snprintf(buffer, sizeof(buffer), "%s %" PRIu32, Instructions::Mnemonic(opcode), (uint32_t)lineNumber);
     }
 
     return (string(buffer));
